cesar_assistant/supabase_flask_app.py
import os
from flask import Flask
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

supabase: Client = create_client(
    os.environ.get("SUPABASE_URL"),
    os.environ.get("SUPABASE_KEY")
)


@app.route('/')
def index():
    try:
        response = supabase.table('todos').select("*").execute()
        todos = response.data

        if not todos:
            return '<h1>No hay tareas pendientes</h1>'

        html = '<h1>Lista de Tareas (Todos)</h1><ul>'
        for todo in todos:
            # Manejo preventivo si la columna 'name' no existe o es nula
            name = todo.get("name", "Sin nombre")
            html += f'<li>{name}</li>'
        html += '</ul>'
        return html
    except Exception as e:
        return f'<h1>Error al conectar con Supabase</h1><p>{str(e)}</p>', 500


if __name__ == '__main__':
    app.run(debug=True)